refine config, add crosspkg tests, add tests for calls
