+++ internal/pkg/config/testdata/src/example.com/notcore/notcore.go
@@ -23,3 +23,11 @@ type Sinker struct{}
 func (s Sinker) Do() {}
 
 func (s Sinker) DoNot() {}
+
+func Calls() {
+	Sink()
+	NotSink()
+	s := Sinker{}
+	s.Do()
+	s.DoNot()
+}
