+++ internal/pkg/config/testdata/test-config.json
@@ -1,11 +1,11 @@
 {
   "Sinks": [
     {
-      "PackageRE": "^core$",
+      "PackageRE": "^example.com/core$",
       "MethodRE": "^Sink$"
     },
     {
-      "PackageRE": "^core$",
+      "PackageRE": "^example.com/core$",
       "MethodRE": "^Do$",
       "TypeRE": "^Sinker$"
     }
